# mykid-service/mykid-rest-adapter/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: mykid
  description: "a funny application that groups together stats for your baby"
  termsOfService: todo
  contact:
    email: todo@todo.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.1-SNAPSHOT
tags:
  - name: my-kid
    description: my kid

paths:

  /v1/stats:

    get:
      operationId: getStats
      tags:
        - my-kid
      parameters:
        - name: startDate
          in: query
          required: false
          description: the start date for the stats
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          description: the end date for the stats
          schema:
            type: string
      responses:
        '200':
          description: all the stats available in the given interval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStatsResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: postStat
      tags:
        - my-kid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Stats'
      responses:
        '200':
          description: all the stats available in the given interval
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


  /v1/auth/signup:
    post:
      tags:
        - user
      summary: Create a user
      description: Create a user
      operationId: signup
      requestBody:
        description: Create a user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-codegen-request-body-name: body

  /v1/auth/login:
    post:
      tags:
        - user
      summary: Use it to login
      description: Use it to login
      operationId: login
      requestBody:
        description: Use it to login
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-codegen-request-body-name: body

components:
  schemas:

    GetStatsResponse:
      type: object
      properties:
        stats:
          type: array
          items:
            $ref: '#/components/schemas/Stats'

    Stats:
      type: object
      required:
        - createdAt
        - datetime
      properties:
        datetime:
          type: string
        activities:
          type: array
          items:
            $ref: '#/components/schemas/GenericActivities'
        temperature:
          type: number
          format: double
        feedFromLeftDuration:
          type: number
          format: double
        feedFromRightDuration:
          type: number
          format: double
        pumpFromLeftQuantity:
          type: number
          format: double
        pumpFromRightQuantity:
          type: number
          format: double
        extraBottleMotherMilkQuantity:
          type: number
          format: double
        extraBottleFormulaeMilkQuantity:
          type: number
          format: double
        weight:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time

    GenericActivities:
      type: string
      enum:
        - BREAST_MILK
        - BOTTLE_MOTHER_MILK
        - BOTTLE_FORMULAE_MILK
        - WEIGHT
        - PUMP
        - SPIT
        - URINE
        - POOP
        - VITAMIN_D
        - VITAMIN_K
        - TEMPERATURE

    User:
      type: object
      required:
        - username
      properties:
        username:
          type: string
        password:
          type: string
        createdAt:
          type: string
          format: date-time
        jwtToken:
          type: string

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string